Overhauled components
I realised it's better if each question/answer is its own object with:
- Question string
- Answer string
- Possible Answers array
- Selected boolean

This means that I can click buttons to select/deselect them. Then, when answers are checked, I can check the selected answer against the correct answer in that object.

Also, it didn't make sense the `Loading` state was in the same object as `triviaQuestions`. This has been updated, along with a `checkingAnswers` state added for future use.

Additionally, now when the API data is fetched, it is instantly parsed and all the relevant data taken in the effect hook, to prevent multiple app refreshes to parse through the data.

Finally, the shuffle() function has been moved to the Menu.js component because of the above. It's more convenient to shuffle the possible answers while making the initial triviaQuestions array/objects, rather than further down the line